requirements.txt
fastapi[all]==0.111.0
pydantic-redis==0.5.0
pydantic==2.7.1
uvicorn==0.29.0
python-dotenv==1.0.1
fuzzywuzzy[speedup]==0.18.0
requests==2.32.2
 beautifulsoup4==4.12.3
 html5lib==1.1